feat: create artifactory-tools package and move script from dev-tools into it
