+++ packages/@ama-sdk/generator-sdk/src/generators/shell/templates/base/__empty__package.json
@@ -123,5 +123,5 @@
   "engines": {
     "node": ">=18.0.0"
   },
-  "packageManager": "yarn@3.6.3"
+  "packageManager": "yarn@3.6.4"
 }
