+++ packages/@o3r/artifactory-tools/.compodocrc.json
@@ -1,0 +1,22 @@
+{
+  "$schema": "https://raw.githubusercontent.com/compodoc/compodoc/develop/src/config/schema.json",
+  "name": "Artifactory tools",
+  "output": "../../../generated-doc/artifactory-tools",
+  "tsconfig": "./tsconfig.doc.json",
+  "assetsFolder": "../../../.attachments",
+  "disableSourceCode": true,
+  "disableDomTree": true,
+  "disableTemplateTab": true,
+  "disableStyleTab": true,
+  "disableGraph": true,
+  "disableCoverage": true,
+  "disablePrivate": true,
+  "disableProtected": true,
+  "disableInternal": true,
+  "disableLifeCycleHooks": true,
+  "disableRoutesGraph": true,
+  "disableSearch": false,
+  "hideGenerator": true,
+  "customFavicon": "../../../.attachments/logos/otter-128x128.png",
+  "templates": "../../../compodoc-templates/package"
+}
