+++ packages/@o3r/artifactory-tools/.eslintrc.js
@@ -1,0 +1,19 @@
+/* eslint-disable @typescript-eslint/naming-convention */
+/* eslint-disable quote-props */
+
+module.exports = {
+  'root': true,
+  'parserOptions': {
+    'tsconfigRootDir': __dirname,
+    'project': [
+      'tsconfig.eslint.json',
+      'tsconfig.build.json',
+      'tsconfig.builders.json',
+      'tsconfig.spec.json'
+    ],
+    'sourceType': 'module'
+  },
+  'extends': [
+    '../../../.eslintrc.js'
+  ]
+};
