+++ packages/@o3r/artifactory-tools/.gitignore
@@ -1,0 +1,2 @@
+build/
+dist/
