+++ packages/@o3r/artifactory-tools/jest.config.js
@@ -1,0 +1,12 @@
+const getJestConfig = require('../../../jest.config.ut').getJestConfig;
+
+/** @type {import('ts-jest/dist/types').JestConfigWithTsJest} */
+module.exports = {
+  ...getJestConfig(__dirname, false),
+  displayName: require('./package.json').name,
+  testPathIgnorePatterns: [
+    '<rootDir>/dist',
+    '<rootDir>/.*/templates/.*',
+    '\\.it\\.spec\\.ts$'
+  ]
+};
