+++ packages/@o3r/artifactory-tools/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "@o3r/artifactory-tools",
+  "version": "0.0.0-placeholder",
+  "description": "Various artifactory tools",
+  "keywords": [
+    "otter",
+    "otter-module"
+  ],
+  "scripts": {
+    "nx": "nx",
+    "ng": "yarn nx",
+    "build": "yarn nx build artifactory-tools",
+    "postbuild": "patch-package-json-main",
+    "prepare:publish": "prepare-publish ./dist",
+    "prepare:build:builders": "yarn cpy 'schematics/**/*.json' dist/schematics && yarn cpy 'collection.json' dist",
+    "build:builders": "tsc -b tsconfig.builders.json --pretty && yarn generate-cjs-manifest"
+  },
+  "bin": {
+    "o3r-artifact-cleaner": "./dist/src/cli/artifact-cleaner.js",
+    "o3r-artifact-retriever": "./dist/src/cli/artifact-retriever.js",
+    "o3r-pr-artifact-cleaner": "./dist/src/cli/pr-artifact-cleaner.js"
+  },
+  "dependencies": {
+    "commander": "^11.0.0",
+    "fs-extra": "^11.0.0",
+    "request": "^2.88.0",
+    "request-promise-native": "^1.0.9",
+    "tslib": "^2.5.3",
+    "winston": "^3.8.2"
+  },
+  "devDependencies": {
+    "@angular-devkit/core": "~16.2.0",
+    "@angular-devkit/schematics": "~16.2.0",
+    "@angular-eslint/eslint-plugin": "~16.1.0",
+    "@babel/core": "~7.22.0",
+    "@babel/preset-typescript": "~7.22.0",
+    "@compodoc/compodoc": "^1.1.19",
+    "@nx/jest": "~16.8.0",
+    "@o3r/build-helpers": "workspace:^",
+    "@o3r/eslint-plugin": "workspace:^",
+    "@types/fs-extra": "^11.0.0",
+    "@types/jest": "~29.5.2",
+    "@types/node": "^18.0.0",
+    "@types/request": "^2.48.7",
+    "@types/request-promise-native": "^1.0.18",
+    "@typescript-eslint/eslint-plugin": "^5.59.9",
+    "@typescript-eslint/parser": "^5.59.9",
+    "cpy-cli": "^4.2.0",
+    "eslint": "^8.42.0",
+    "eslint-import-resolver-node": "^0.3.4",
+    "eslint-plugin-jest": "~27.2.3",
+    "eslint-plugin-jsdoc": "~46.8.0",
+    "eslint-plugin-prefer-arrow": "~1.2.3",
+    "eslint-plugin-unicorn": "^47.0.0",
+    "jest": "~29.7.0",
+    "jest-junit": "~16.0.0",
+    "nx": "~16.8.0",
+    "ts-jest": "~29.1.1",
+    "typescript": "~5.1.6"
+  },
+  "engines": {
+    "node": ">=16.0.0"
+  },
+  "schematics": "./collection.json"
+}
