+++ packages/@o3r/artifactory-tools/project.json
@@ -1,0 +1,87 @@
+{
+  "name": "artifactory-tools",
+  "$schema": "https://raw.githubusercontent.com/nrwl/nx/master/packages/nx/schemas/project-schema.json",
+  "sourceRoot": "packages/@o3r/artifactory-tools/src",
+  "prefix": "o3r",
+  "tags": ["tools"],
+  "projectType": "library",
+  "targets": {
+    "build": {
+      "executor": "nx:run-script",
+      "outputs": [
+        "{projectRoot}/dist/package.json"
+      ],
+      "options": {
+        "script": "postbuild"
+      },
+      "dependsOn": [
+        "compile",
+        "build-builders"
+      ]
+    },
+    "prepare-build-builders": {
+      "executor": "nx:run-script",
+      "options": {
+        "script": "prepare:build:builders"
+      }
+    },
+    "build-builders": {
+      "executor": "nx:run-script",
+      "options": {
+        "script": "build:builders"
+      }
+    },
+    "compile": {
+      "executor": "@nx/js:tsc",
+      "options": {
+        "main": "packages/@o3r/artifactory-tools/src/public_api.ts",
+        "tsConfig": "packages/@o3r/artifactory-tools/tsconfig.build.json",
+        "outputPath": "packages/@o3r/artifactory-tools/dist",
+        "assets": [
+          "packages/@o3r/artifactory-tools/configs/*.json"
+        ],
+        "clean": false
+      }
+    },
+    "lint": {
+      "executor": "@nx/linter:eslint",
+      "configurations": {
+        "ci": {
+          "quiet": true,
+          "cacheLocation": ".cache/eslint"
+        }
+      },
+      "options": {
+        "eslintConfig": "packages/@o3r/artifactory-tools/.eslintrc.js",
+        "lintFilePatterns": [
+          "packages/@o3r/artifactory-tools/src/**/*.ts"
+        ]
+      }
+    },
+    "test": {
+      "executor": "@nx/jest:jest",
+      "options": {
+        "jestConfig": "packages/@o3r/artifactory-tools/jest.config.js",
+        "silent": true
+      }
+    },
+    "prepare-publish": {
+      "executor": "nx:run-script",
+      "options": {
+        "script": "prepare:publish"
+      }
+    },
+    "publish": {
+      "executor": "nx:run-commands",
+      "options": {
+        "command": "npm publish packages/@o3r/artifactory-tools/dist"
+      }
+    },
+    "documentation": {
+      "executor": "nx:run-script",
+      "options": {
+        "script": "compodoc"
+      }
+    }
+  }
+}
