+++ packages/@o3r/artifactory-tools/schematics/ng-add/index.ts
@@ -1,0 +1,12 @@
+import { noop } from '@angular-devkit/schematics';
+import type { Rule } from '@angular-devkit/schematics';
+
+/**
+ * Add Otter artifactory-tools to an Angular Project
+ *
+ * @param options
+ */
+export function ngAdd(): Rule {
+  /* ng add rules */
+  return noop();
+}
