+++ packages/@o3r/artifactory-tools/schematics/ng-add/schema.ts
@@ -1,0 +1,4 @@
+export interface NgAddSchematicsSchema {
+  /** Project name */
+  projectName?: string | undefined;
+}
