+++ packages/@o3r/artifactory-tools/src/index.spec.ts
@@ -1,0 +1,5 @@
+// TODO be removed as soon as we have one test in this package
+
+it('should be removed as soon as we have one test in this package', () => {
+  expect(true).toBe(true);
+});
