+++ packages/@o3r/artifactory-tools/src/public_api.ts
@@ -1,0 +1,1 @@
+export type {};
