+++ packages/@o3r/artifactory-tools/tsconfig.eslint.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.build",
+  "include": [
+    ".eslintrc.js",
+    "jest.config.js",
+    "testing/*"
+  ]
+}
