+++ packages/@o3r/artifactory-tools/tsconfig.spec.json
@@ -1,0 +1,16 @@
+{
+  "extends": "../../../tsconfig.jest",
+  "compilerOptions": {
+    "outDir": "test",
+    "rootDir": "."
+  },
+  "include": [
+    "./src/**/*.spec.ts"
+  ],
+  "exclude": [],
+  "references": [
+    {
+      "path": "./tsconfig.build.json"
+    }
+  ]
+}
